// src/shared/types/offer.type.ts
import { City } from './city.type.js';
import { OfferType } from './offer-type.enum.js';
import { Comfort } from './comfort.enum.js';
import { User } from './user.type.js';
import { Coordinates } from './coordinates.type.js';

export type Offer = {
  // Наименование. Обязательное. Мин. длин 10 символов, макс. длина 100;
  title: string;
  // Описание предложения. Обязательное. Мин. длина 20 символов, макс. длина 1024 символа;
  description: string;
  // Дата публикации предложения. Обязательное.
  publicationTime: string; // ISO
  // Город. Обязательное. Один из шести городов.
  city: City;
  // Превью изображения. Обязательное. Ссылка на изображение, которое используется в качестве превью;
  previewPic: string;
  // Фотографии жилья. Обязательное. Список ссылок на фотографии жилья. Всегда 6 фотографий;
  photos: string[];
  // Флаг «Премиум». Обязательное. Признак премиальности предложения;
  isPremium: boolean;
  // Флаг «Избранное». Обязательное. Признак того, что предложение принадлежит списку избранных предложений пользователя;
  isBookmarked: boolean;
  // Рейтинг. Обязательное. Число от 1 до 5. Допускаются числа с запятой (1 знак после запятой);
  rating: number;
  // Тип жилья. Обязательное. Один из вариантов: apartment, house, room, hotel;
  type: OfferType;
  // Количество комнат. Обязательное. Мин. 1, Макс. 8;
  rooms: number;
  // Количество гостей. Обязательное. Мин. 1, Макс. 10;
  guests: number;
  // Стоимость аренды. Обязательное. Мин. 100, Макс. 100 000;
  price: number;
  // Удобства. Обязательное. Список удобств.
  comforts: Comfort[];
  // Автор предложения. Обязательное. Ссылка на сущность «Пользователь»;
  user: User;
  // Количество комментариев. Рассчитывается автоматически;
  // Координаты предложения для аренды. Обязательное. Координаты представлены широтой и долготой.
  coordinates: Coordinates;
}